Print statment Improved

+++ ValueHandler.cs
@@ -43,7 +43,7 @@ namespace Gnosis
             string expression = string.Join("",tokens);
             var rt = MathEngine.ReturnType.Double;
 
-            return ParseNumberExpression(expression,rt);
+            return (double)ParseNumberExpression(expression,rt);
             //var value = blah + blah + blah
         }
 
@@ -55,7 +55,7 @@ namespace Gnosis
             string expression = string.Join("", tokens);
             var rt = MathEngine.ReturnType.Float;
 
-            return ParseNumberExpression(expression, rt);
+            return (float)ParseNumberExpression(expression, rt);
             //var value = blah + blah + blah
         }
 
@@ -79,14 +79,18 @@ namespace Gnosis
             string expression = string.Join("", tokens);
             var rt = MathEngine.ReturnType.Long;
 
-            return ParseNumberExpression(expression, rt);
+            return (long)ParseNumberExpression(expression, rt);
             //var value = blah + blah + blah
         }
 
-        public string ParseStringExpression(List<string> tokens)
+        public string ParseStringExpression(List<string> tokens, bool fromVariable = true)
         {
-            tokens.RemoveRange(0, 3);
-            tokens.RemoveAt(tokens.Count - 1);
+            if(fromVariable)
+            {
+                tokens.RemoveRange(0, 3);
+                tokens.RemoveAt(tokens.Count - 1);
+            }
+            
 
             string output = "";
 
@@ -101,6 +105,32 @@ namespace Gnosis
                 }
                 else
                 {
+                    output += next;
+                }
+            }
+
+            return output;
+            //var value = blah + blah + blah
+        }
+
+        public string ParseBoolExpression(List<string> tokens)
+        {
+            tokens.RemoveRange(0, 3);
+            tokens.RemoveAt(tokens.Count - 1);
+
+            string output = "";
+
+            for (int i = 0; i < tokens.Count; i++)
+            {
+                var next = GetValue(tokens[i]);
+
+                if (next == null)
+                {
+                    //Throw exception
+                    //To be implemented
+                }
+                else
+                {
                     output += GetValue(tokens[i]);
                 }
             }
@@ -109,11 +139,19 @@ namespace Gnosis
             //var value = blah + blah + blah
         }
 
+
         public dynamic GetValue(string value)
         {
             if (Lexer.IsString(value))
             {
-                return value.Substring(1, value.Length - 2);
+                return value.Substring(1, value.Length - 2)
+                    .Replace("\\a", "\a")
+                    .Replace("\\b", "\b")
+                    .Replace("\\f", "\f")
+                    .Replace("\\n", "\n")
+                    .Replace("\\r", "\r")
+                    .Replace("\\t", "\t")
+                    .Replace("\\v", "\v");
             }
             else if (method.lexer.isVariable(value))
             {
